requirements.txt
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
mysqlclient==2.2.4
netifaces==0.11.0
nltk==3.8.1
pipenv==2024.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
virtualenv==20.26.3
whitenoise>=6.0.0,<7.0.0
